Revert "Update master to v3.7.0"
This reverts commit fbbd582ee1bb5435c78998f471bdd29663feccc0.
