+++ callingconv_gen.go
@@ -9,20 +9,20 @@ import "fmt"
 type CallingConv uint32
 
 const (
-	CallingConv_Default       CallingConv = C.CXCallingConv_Default
-	CallingConv_C                         = C.CXCallingConv_C
-	CallingConv_X86StdCall                = C.CXCallingConv_X86StdCall
-	CallingConv_X86FastCall               = C.CXCallingConv_X86FastCall
-	CallingConv_X86ThisCall               = C.CXCallingConv_X86ThisCall
-	CallingConv_X86Pascal                 = C.CXCallingConv_X86Pascal
-	CallingConv_AAPCS                     = C.CXCallingConv_AAPCS
-	CallingConv_AAPCS_VFP                 = C.CXCallingConv_AAPCS_VFP
-	CallingConv_IntelOclBicc              = C.CXCallingConv_IntelOclBicc
-	CallingConv_X86_64Win64               = C.CXCallingConv_X86_64Win64
-	CallingConv_X86_64SysV                = C.CXCallingConv_X86_64SysV
-	CallingConv_X86VectorCall             = C.CXCallingConv_X86VectorCall
-	CallingConv_Invalid                   = C.CXCallingConv_Invalid
-	CallingConv_Unexposed                 = C.CXCallingConv_Unexposed
+	CallingConv_Default      CallingConv = C.CXCallingConv_Default
+	CallingConv_C                        = C.CXCallingConv_C
+	CallingConv_X86StdCall               = C.CXCallingConv_X86StdCall
+	CallingConv_X86FastCall              = C.CXCallingConv_X86FastCall
+	CallingConv_X86ThisCall              = C.CXCallingConv_X86ThisCall
+	CallingConv_X86Pascal                = C.CXCallingConv_X86Pascal
+	CallingConv_AAPCS                    = C.CXCallingConv_AAPCS
+	CallingConv_AAPCS_VFP                = C.CXCallingConv_AAPCS_VFP
+	CallingConv_PnaclCall                = C.CXCallingConv_PnaclCall
+	CallingConv_IntelOclBicc             = C.CXCallingConv_IntelOclBicc
+	CallingConv_X86_64Win64              = C.CXCallingConv_X86_64Win64
+	CallingConv_X86_64SysV               = C.CXCallingConv_X86_64SysV
+	CallingConv_Invalid                  = C.CXCallingConv_Invalid
+	CallingConv_Unexposed                = C.CXCallingConv_Unexposed
 )
 
 func (cc CallingConv) Spelling() string {
@@ -43,14 +43,14 @@ func (cc CallingConv) Spelling() string {
 		return "CallingConv=AAPCS"
 	case CallingConv_AAPCS_VFP:
 		return "CallingConv=AAPCS_VFP"
+	case CallingConv_PnaclCall:
+		return "CallingConv=PnaclCall"
 	case CallingConv_IntelOclBicc:
 		return "CallingConv=IntelOclBicc"
 	case CallingConv_X86_64Win64:
 		return "CallingConv=X86_64Win64"
 	case CallingConv_X86_64SysV:
 		return "CallingConv=X86_64SysV"
-	case CallingConv_X86VectorCall:
-		return "CallingConv=X86VectorCall"
 	case CallingConv_Invalid:
 		return "CallingConv=Invalid"
 	case CallingConv_Unexposed:
