+++ clang-c/CXCompilationDatabase.h
@@ -14,8 +14,8 @@
 |*                                                                            *|
 \*===----------------------------------------------------------------------===*/
 
-#ifndef LLVM_CLANG_C_CXCOMPILATIONDATABASE_H
-#define LLVM_CLANG_C_CXCOMPILATIONDATABASE_H
+#ifndef CLANG_CXCOMPILATIONDATABASE_H
+#define CLANG_CXCOMPILATIONDATABASE_H
 
 #include "clang-c/Platform.h"
 #include "clang-c/CXString.h"
