+++ clang-c/CXString.h
@@ -13,8 +13,8 @@
 |*                                                                            *|
 \*===----------------------------------------------------------------------===*/
 
-#ifndef LLVM_CLANG_C_CXSTRING_H
-#define LLVM_CLANG_C_CXSTRING_H
+#ifndef CLANG_CXSTRING_H
+#define CLANG_CXSTRING_H
 
 #include "clang-c/Platform.h"
 
@@ -33,7 +33,7 @@ extern "C" {
  * \brief A character string.
  *
  * The \c CXString type is used to return strings from the interface when
- * the ownership of that string might differ from one call to the next.
+ * the ownership of that string might different from one call to the next.
  * Use \c clang_getCString() to retrieve the string data and, once finished
  * with the string data, call \c clang_disposeString() to free the string.
  */
