+++ clang-c/Platform.h
@@ -13,8 +13,8 @@
 |*                                                                            *|
 \*===----------------------------------------------------------------------===*/
 
-#ifndef LLVM_CLANG_C_PLATFORM_H
-#define LLVM_CLANG_C_PLATFORM_H
+#ifndef CLANG_C_PLATFORM_H
+#define CLANG_C_PLATFORM_H
 
 #ifdef __cplusplus
 extern "C" {
