--- clang-c/module.modulemap
@@ -1,4 +1,0 @@
-module Clang_C {
-  umbrella "."
-  module * { export * }
-}
