+++ comment_gen.go
@@ -1,10 +1,10 @@
 package phoenix
 
-// #include "./clang-c/Documentation.h"
+// #include "./clang-c/Index.h"
 // #include "go-clang.h"
 import "C"
 
-// A parsed comment.
+// A comment AST node.
 type Comment struct {
 	c C.CXComment
 }
