+++ declqualifierkind_gen.go
@@ -5,7 +5,7 @@ package phoenix
 import "C"
 import "fmt"
 
-// 'Qualifiers' written next to the return and parameter types in Objective-C method declarations.
+// 'Qualifiers' written next to the return and parameter types in ObjC method declarations.
 type DeclQualifierKind uint32
 
 const (
