+++ string_gen.go
@@ -8,7 +8,7 @@ import "C"
 	A character string.
 
 	The CXString type is used to return strings from the interface when
-	the ownership of that string might differ from one call to the next.
+	the ownership of that string might different from one call to the next.
 	Use clang_getCString() to retrieve the string data and, once finished
 	with the string data, call clang_disposeString() to free the string.
 */
